Add support for introspective generated tests
The basic idea is that instead of having to provide generators for
parameters, the test runner will generate those based on metadata
(_invariants). It will use a generator (suite._generator) to find
values filling to those invariants and then uses these as tests.

Compared to previous solution the syntax in this is a lot lighter
and moves part of the invariants to the right place (DbC).